+++ scripts/__init__.py
@@ -1,0 +1,1 @@
+"""Support code for the reusable GitHub Action."""